[cuBLAS][rocBLAS] Fix uninstantiated template static asserts for older compilers
CWG2518 allowing static_assert(false) in uninstantiated template is a C++23
behaviour that happens to work regardless of -std flag in recent compilers
(clang >= 17) including DPC++. However, AdaptiveCpp is based on an older clang
version and does not support that.

Force the static assert evaluation to be pushed after template evaluations
by using type T inside the expression. This works in all supported compilers.

Also add a message in the assert to clarify why it's there.