[BLAS][generic] Fix compilation with AdaptiveCpp
The generic BLAS backend currently offers limited support for
AdaptiveCpp, where:
* complex data type is not supported
* USM API is not supported

Add the required protections to make the generic BLAS backend
compile and run correctly in the capacity it offers with AdaptiveCpp.
That is, make the buffer USM without complex data work fine. Throw
the unimplemented exception for the unsupported features.

This change relies on the update of the generic blas backend to
version 0.2.0 in:
https://github.com/uxlfoundation/generic-sycl-components/pull/7