Fix preprocessor macros protecting DPC++ bfloat16 extension
Use the SYCL2020-standard macros `SYCL_IMPLEMENTATION_<vendor>` to
decide whether the extension is available or not.

Previous version assumed that if the compiler is not AdaptiveCpp,
then it is DPC++. Although these are only currently supported
compilers, such assumptions block the development to support other
SYCL implementations. There is also no good reason to block non-SYCL
compilers from parsing the oneMath headers in such a way.

Ideally, the use of an extension should be protected by a specific
feature-test macro, but this is currently missing for the DPC++
bfloat16 extension.

Note DPC++ provides two vendor identifiers `SYCL_IMPLEMENTATION_ONEAPI`
and `SYCL_IMPLEMENTATION_INTEL`. The latter is used here to avoid
confusion with the name "oneAPI" referring to "oneAPI Math Library"
which could mean an internal implementation of a feature in this project.