[LAPACK][tests] Fix is_buf helper for non-DPC++ compilers
The SYCL standard `buffer` class template has three template arguments.
DPC++ uses a fourth argument with a default value depending on the
second argument in order to prevent compiling code with dimension
different than 1, 2 or 3. This fourth argument is an implementation
detail of DPC++ and SYCL code should not assume its existence.

The is_buf helper works exactly the same without the fourth argument
and aligning it with the SYCL standard fixes a compilation failure
with AdaptiveCpp:
error: too many template arguments for class template 'buffer'